Phishing-Detection-Extension/requirements.txt
flask
flask-cors
flask-sqlalchemy
joblib
numpy
scipy
requests
scikit-learn
xgboost
gunicorn
python-whois